Move files to pkg dir
