+++ tools/codegen/pkg/modelgen/Diag.go
