+++ tools/codegen/pkg/modelgen/parser.go
