+++ tools/codegen/pkg/modelgen/testdata/BasicTopLevelFields.proto
