+++ tools/codegen/pkg/modelgen/testdata/ConstructorAndImports.proto
