+++ tools/codegen/pkg/modelgen/testdata/MissingBothRequiredExt.proto
