+++ tools/codegen/pkg/modelgen/testdata/MissingConstructorMessage.proto
