+++ tools/codegen/pkg/modelgen/testdata/MissingPackageName.proto
