+++ tools/codegen/pkg/modelgen/testdata/MissingTemplateNameExt.proto
