+++ tools/codegen/pkg/modelgen/testdata/MissingTemplateVarietyExt.proto
