+++ tools/codegen/pkg/modelgen/testdata/MissingTypeMessage.proto
