+++ tools/codegen/pkg/proc_interface_gen/template/ProcInterface.tmpl
