+++ tools/codegen/pkg/proc_interface_gen/testdata/ErrorTemplate.proto
