+++ tools/codegen/pkg/proc_interface_gen/testdata/ListTemplateProcessorInterface.go.baseline
