+++ tools/codegen/pkg/proc_interface_gen/testdata/LogTemplateProcessorInterface.go.baseline
