+++ tools/codegen/pkg/proc_interface_gen/testdata/MetricTemplateProcessorInterface.go.baseline
