+++ tools/codegen/pkg/proc_interface_gen/testdata/NestedMessage.proto
