+++ tools/codegen/pkg/proc_interface_gen/testdata/NestedMessageProcessor.go.baseline
