+++ tools/codegen/pkg/proc_interface_gen/testdata/QuotaTemplateProcessorInterface.go.baseline
